refactor: rewrote the main to classes
