+++ src/assets/scripts/types/Interfaces.ts
@@ -1,4 +1,4 @@
-interface book {
+export interface book {
   id: number,
   publisher: string,
   description: string,
@@ -10,4 +10,12 @@ interface book {
   stock_balance: number,
 }
 
-export {book}
+export function getElementBySelector <T extends typeof Element>(
+  element: DocumentFragment | HTMLElement | Document,
+  type: T,
+  selector: string): InstanceType<T> {
+  const result = element.querySelector(selector);
+  if (!result) throw new Error(`Selector ${selector} didn't match any elements.`);
+  if (!(result instanceof type)) throw new TypeError(`Selector ${selector} have wrong type`);
+  return result as InstanceType<T>;
+}
