+++ src/index.html
@@ -25,48 +25,10 @@
     </div>
   </header>
   <section class="filters">
-    <div class="container filters__container">
-      <div class="filters__wrapper">
-        <div class="filters__genre">
-          <h3 class="filters__genre-title">Genre</h3>
-        </div>
-        <div class="filters__author">
-          <h3 class="filters__authors-title">Author</h3>
-        </div>
-        <div class="filters__sliders">
-          <div class="filters__price">
-            <h3 class="filters__price-title">Price</h3>
-            <div class="filters__price-minmax">
-              <p class="filters__price-min">0</p>
-              ⟷
-              <p class="filters__price-max">0</p>
-            </div>
-            <div id="slider-price"></div>
-          </div>
-          <div class="filters__stock">
-            <h3 class="filters__stock-title">Stock</h3>
-            <div class="filters__stock-minmax">
-              <p class="filters__stock-min">0</p>
-              ⟷
-              <p class="filters__stock-max">0</p>
-            </div>
-            <div id="slider-stock"></div>
-          </div>
-        </div>
-      </div>
-      <button class="filters__filter-reset">Reset Filters</button>
-      <div class="filters__search">
-        <form class="filters__search-form">
-          <input class="filters__search-input" type="search" name="q" placeholder="Search for books by keyword">
-          <input class="filters__search-submit" type="submit" value="">
-        </form>
-      </div>
-    </div>
+    <div class="container filters__container"></div>
   </section>
   <main class="main">
-    <div class="container main__container">
-
-    </div>
+    <div class="container main__container"></div>
   </main>
   <footer class="footer">
     <div class="container footer__container">
